% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_data_fn_nor_covid19_cases_by_time_location}
\alias{example_data_fn_nor_covid19_cases_by_time_location}
\title{Return the example Covid-19 dataset}
\usage{
example_data_fn_nor_covid19_cases_by_time_location()
}
\value{
\link{nor_covid19_cases_by_time_location}, a data.table.
}
\description{
\code{example_data_fn_nor_covid19_cases_by_time_location()} is an example data
function: it takes no arguments and returns one dataset. Give its name to
\code{Plan$add_data()} as \code{fn_name}.
}
\examples{
dim(example_data_fn_nor_covid19_cases_by_time_location())

p <- plnr::Plan$new()
p$add_data(
  name = "covid19_cases",
  fn_name = "plnr::example_data_fn_nor_covid19_cases_by_time_location"
)
names(p$get_data())
}
\seealso{
\code{vignette("adding_analyses")}, which writes its own data functions.

Other example and test functions:
\code{\link[=example_action_fn]{example_action_fn()}},
\code{\link[=test_action_fn]{test_action_fn()}}
}
\concept{example and test functions}
