% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{hash_it}
\alias{hash_it}
\title{Hash an object}
\usage{
hash_it(x)
}
\arguments{
\item{x}{The object to hash.}
}
\value{
A character string that holds the hash.
}
\description{
\code{hash_it()} is internal. It returns \code{digest::digest(x)}.
}
\keyword{internal}
