% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nor_covid19_cases_by_time_location}
\alias{nor_covid19_cases_by_time_location}
\title{Covid-19 cases confirmed by PCR in Norway, by nation and county}
\format{
A data.table with 11028 rows and 18 variables:
\describe{
\item{granularity_time}{\code{"day"} or \code{"isoweek"}.}
\item{granularity_geo}{\code{"nation"} or \code{"county"}.}
\item{country_iso3}{\code{"nor"}.}
\item{location_code}{\code{"nation_nor"}, or one of 11 county codes such as \code{"county_nor03"}.}
\item{border}{The year of the county borders, 2020.}
\item{age}{\code{"total"}.}
\item{sex}{\code{"total"}.}
\item{isoyear}{ISO year.}
\item{isoweek}{ISO week.}
\item{isoyearweek}{ISO year and week, such as \code{"2020-08"}.}
\item{season}{Season, such as \code{"2020/2021"}.}
\item{seasonweek}{Week within the season.}
\item{calyear}{Calendar year. \code{NA} in weekly rows.}
\item{calmonth}{Calendar month. \code{NA} in weekly rows.}
\item{calyearmonth}{Calendar year and month, such as \code{"2020-M02"}. \code{NA} in weekly rows.}
\item{date}{The date. In weekly rows, the Sunday that ends the ISO week.}
\item{covid19_cases_testdate_n}{Number of confirmed cases.}
\item{covid19_cases_testdate_pr100000}{Number of confirmed cases per 100 000 population.}
}
}
\source{
\url{https://github.com/folkehelseinstituttet/surveillance_data/blob/master/covid19/_DOCUMENTATION_data_covid19_msis_by_time_location.txt}
}
\usage{
nor_covid19_cases_by_time_location
}
\description{
The data come from the Norwegian Surveillance System for Communicable
Diseases (MSIS). \code{date} is the date of the PCR test. The data were extracted
on 2022-05-04.
}
\keyword{datasets}
