% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\docType{data}
\name{PROJ}
\alias{PROJ}
\title{Folders to be used/referenced}
\format{An object of class \code{environment} of length 0.}
\usage{
PROJ
}
\description{
Folders to be used/referenced
}
\keyword{datasets}
