% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_opts.R
\name{set_opts}
\alias{set_opts}
\title{Set plnr options}
\usage{
set_opts(force_verbose = FALSE)
}
\arguments{
\item{force_verbose}{Logical. \code{TRUE} makes every \link{Plan} that you create
afterwards verbose, even in a non-interactive session. The default is
\code{FALSE}.}
}
\value{
\code{force_verbose}, invisibly.
}
\description{
\code{set_opts()} sets options for the whole package.
}
\examples{
set_opts(force_verbose = TRUE)
set_opts(force_verbose = FALSE)
}
\seealso{
\code{vignette("plnr")} for the concepts.

Other plan helpers:
\code{\link{Plan}},
\code{\link[=expand_list]{expand_list()}},
\code{\link[=get_anything]{get_anything()}},
\code{\link[=is_run_directly]{is_run_directly()}}
}
\concept{plan helpers}
