% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{test_action_fn}
\alias{test_action_fn}
\title{Return 1, whatever the input}
\usage{
test_action_fn(data, argset)
}
\arguments{
\item{data}{Not used.}

\item{argset}{Not used.}
}
\value{
The number 1, a double.
}
\description{
\code{test_action_fn()} is a placeholder action function. It ignores its arguments.
}
\examples{
test_action_fn(data = list(), argset = list())

p <- plnr::Plan$new()
p$add_data(name = "deaths", direct = data.frame(deaths = 1:4, year = 2001:2004))
p$add_analysis(name = "analysis_1", fn_name = "plnr::test_action_fn")
p$run_one("analysis_1")
}
\seealso{
\code{vignette("plnr")} for what an action function is.

Other example and test functions:
\code{\link[=example_action_fn]{example_action_fn()}},
\code{\link[=example_data_fn_nor_covid19_cases_by_time_location]{example_data_fn_nor_covid19_cases_by_time_location()}}
}
\concept{example and test functions}
