% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_progress_bar.R
\name{txt_progress_bar}
\alias{txt_progress_bar}
\title{txt_progress_bar}
\usage{
txt_progress_bar(min = 0, max = 1, initial = 0, char = "=",
  width = NA, title, label, style = 1, file = "",
  time_start = Sys.time())
}
\arguments{
\item{min}{a}

\item{max}{a}

\item{initial}{a}

\item{char}{a}

\item{width}{a}

\item{title}{a}

\item{label}{a}

\item{style}{a}

\item{file}{a}

\item{time_start}{a}
}
\description{
txt_progress_bar
}
