% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rmarkdown.R
\name{create_rmarkdown}
\alias{create_rmarkdown}
\title{Create rmarkdown project}
\usage{
create_rmarkdown(home)
}
\arguments{
\item{home}{Location of the 'home' directory}
}
\description{
Create rmarkdown project
}
