% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_equal.r
\name{easy_split}
\alias{easy_split}
\title{easy_split}
\usage{
easy_split(x, size_of_each_group = NULL, number_of_groups = NULL)
}
\arguments{
\item{x}{The variable to be split}

\item{size_of_each_group}{If you want to split `x` into a number of groups, each of `size_of_each_group` size}

\item{number_of_groups}{How many equally sized groups do you want?}
}
\description{
easy_split
}
\examples{
easy_split(letters[1:20], size_of_each_group = 3)
easy_split(letters[1:20], number_of_groups = 3)
}
