% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{example_action_fn}
\alias{example_action_fn}
\title{An example action_fn for an analysis}
\usage{
example_action_fn(data, argset)
}
\arguments{
\item{data}{Named list.}

\item{argset}{Named list.}
}
\description{
An example action_fn for an analysis
}
