% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{example_action_fn}
\alias{example_action_fn}
\title{An example action_fn for an analysis}
\usage{
example_action_fn(data, argset)
}
\arguments{
\item{data}{Named list.}

\item{argset}{Named list.}
}
\description{
An example action_fn for an analysis
}
\examples{
p <- plnr::Plan$new()
p$add_data("covid_data", fn_name = "plnr::example_data_fn_nor_covid19_cases_by_time_location")
batch_argset_list <- list(
  list(name = "analysis_1", var_1 = 1, var_2 = "i"),
  list(name = "analysis_2", var_1 = 2, var_2 = "j"),
  list(name = "analysis_3", var_1 = 3, var_2 = "k")
)
p$add_analysis_from_list(
  fn_name = "plnr::example_action_fn",
  l = batch_argset_list
)
p$get_argsets_as_dt()
p$run_one("analysis_1")
}
