% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_data_fn_nor_covid19_cases_by_time_location}
\alias{example_data_fn_nor_covid19_cases_by_time_location}
\title{An example data_fn that returns a data set}
\usage{
example_data_fn_nor_covid19_cases_by_time_location()
}
\description{
An example data_fn that returns a data set
}
