% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_data_fn_nor_covid19_cases_by_time_location}
\alias{example_data_fn_nor_covid19_cases_by_time_location}
\title{An example data_fn that returns a data set}
\usage{
example_data_fn_nor_covid19_cases_by_time_location()
}
\description{
An example data_fn that returns a data set
}
\examples{
# A data function takes no arguments and returns one data set
d <- example_data_fn_nor_covid19_cases_by_time_location()
dim(d)

# Its intended use is as an `fn_name` passed to Plan$add_data()
p <- plnr::Plan$new()
p$add_data(
  name = "covid19_cases",
  fn_name = "plnr::example_data_fn_nor_covid19_cases_by_time_location"
)
names(p$get_data())
}
\seealso{
\code{vignette("adding_analyses")}. It defines the same kind of
zero-argument \code{data_fn}. It then attaches that \code{data_fn} to a plan with
\code{fn_name}.

Other example and test functions:
\code{\link[=example_action_fn]{example_action_fn()}},
\code{\link[=test_action_fn]{test_action_fn()}}
}
\concept{example and test functions}
