% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_list.R
\name{expand_list}
\alias{expand_list}
\title{expand_list
The same as purrr::cross, but doesn't require an extra list()}
\usage{
expand_list(...)
}
\arguments{
\item{...}{Dots}
}
\description{
expand_list
The same as purrr::cross, but doesn't require an extra list()
}
