% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_list.R
\name{expand_list}
\alias{expand_list}
\title{Expand (cross) lists}
\usage{
expand_list(...)
}
\arguments{
\item{...}{Dots}
}
\description{
The same as purrr::cross, but doesn't require an extra list()
}
\examples{
plnr::expand_list(
  a = 1:2,
  b = c("a", "b")
)
}
