% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anything.R
\name{get_anything}
\alias{get_anything}
\title{Gets anything (including with package scoping)}
\usage{
get_anything(x)
}
\arguments{
\item{x}{the string that we are getting}
}
\description{
base::get does not work with package scoping (e.g. get("pkg::var")).
plnr::get_anything works with package scoping.
}
\examples{
plnr::get_anything("plnr::nor_covid19_cases_by_time_location")
}
