% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{hash_it}
\alias{hash_it}
\title{Generate a hash of an object}
\usage{
hash_it(x)
}
\arguments{
\item{x}{The object to hash.}
}
\value{
A character string that holds the hash.
}
\description{
\code{hash_it()} is an internal function. It creates a hash of an object with the
digest package.
}
\keyword{internal}
