% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_run_directly.R
\name{is_run_directly}
\alias{is_run_directly}
\title{Is this code run directly?}
\usage{
is_run_directly()
}
\description{
This function determines if it is being called from within a function
or if it is being run directly
}
