% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\docType{data}
\name{Plan}
\alias{Plan}
\title{plan class description}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Plan
}
\description{
An argset is:
- a set of arguments
}
\details{
An analysis is:
- one argset
- one function

A plan is:
- one data pull
- a list of analyses
}
\keyword{datasets}
