% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\docType{data}
\name{Plan}
\alias{Plan}
\title{plan class description}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Plan
}
\description{
A plan is:
- one data pull
- multiple sets of arguments, each of which call a function once
}
\keyword{datasets}
