% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{Plan}
\alias{Plan}
\title{plan class description}
\description{
plan class description

plan class description
}
\details{
An argset is:
\itemize{
\item a set of arguments
}

An analysis is:
\itemize{
\item one argset
\item one function
}

A plan is:
\itemize{
\item one data pull
\item a list of analyses
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Plan$new()}}
\item \href{#method-add_data}{\code{Plan$add_data()}}
\item \href{#method-add_argset}{\code{Plan$add_argset()}}
\item \href{#method-add_argset_from_df}{\code{Plan$add_argset_from_df()}}
\item \href{#method-add_argset_from_list}{\code{Plan$add_argset_from_list()}}
\item \href{#method-add_analysis}{\code{Plan$add_analysis()}}
\item \href{#method-add_analysis_from_df}{\code{Plan$add_analysis_from_df()}}
\item \href{#method-add_analysis_from_list}{\code{Plan$add_analysis_from_list()}}
\item \href{#method-apply_analysis_fn_to_all}{\code{Plan$apply_analysis_fn_to_all()}}
\item \href{#method-len}{\code{Plan$len()}}
\item \href{#method-x_seq_along}{\code{Plan$x_seq_along()}}
\item \href{#method-set_progress}{\code{Plan$set_progress()}}
\item \href{#method-set_progressor}{\code{Plan$set_progressor()}}
\item \href{#method-set_verbose}{\code{Plan$set_verbose()}}
\item \href{#method-get_data}{\code{Plan$get_data()}}
\item \href{#method-get_analysis}{\code{Plan$get_analysis()}}
\item \href{#method-get_argset}{\code{Plan$get_argset()}}
\item \href{#method-run_one_with_data}{\code{Plan$run_one_with_data()}}
\item \href{#method-run_one}{\code{Plan$run_one()}}
\item \href{#method-use_foreach_decision}{\code{Plan$use_foreach_decision()}}
\item \href{#method-run_all_with_data}{\code{Plan$run_all_with_data()}}
\item \href{#method-run_all}{\code{Plan$run_all()}}
\item \href{#method-run_all_progress}{\code{Plan$run_all_progress()}}
\item \href{#method-clone}{\code{Plan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$new(
  argset_name = "argset",
  verbose = interactive() | config$force_verbose,
  use_foreach = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_data"></a>}}
\if{latex}{\out{\hypertarget{method-add_data}{}}}
\subsection{Method \code{add_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$add_data(name, fn = NULL, fn_name = NULL, direct = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_argset"></a>}}
\if{latex}{\out{\hypertarget{method-add_argset}{}}}
\subsection{Method \code{add_argset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$add_argset(name = uuid::UUIDgenerate(), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_argset_from_df"></a>}}
\if{latex}{\out{\hypertarget{method-add_argset_from_df}{}}}
\subsection{Method \code{add_argset_from_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$add_argset_from_df(df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_argset_from_list"></a>}}
\if{latex}{\out{\hypertarget{method-add_argset_from_list}{}}}
\subsection{Method \code{add_argset_from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$add_argset_from_list(l)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-add_analysis}{}}}
\subsection{Method \code{add_analysis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$add_analysis(name = uuid::UUIDgenerate(), fn = NULL, fn_name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_analysis_from_df"></a>}}
\if{latex}{\out{\hypertarget{method-add_analysis_from_df}{}}}
\subsection{Method \code{add_analysis_from_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$add_analysis_from_df(fn = NULL, fn_name = NULL, df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_analysis_from_list"></a>}}
\if{latex}{\out{\hypertarget{method-add_analysis_from_list}{}}}
\subsection{Method \code{add_analysis_from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$add_analysis_from_list(fn = NULL, fn_name = NULL, l)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_analysis_fn_to_all"></a>}}
\if{latex}{\out{\hypertarget{method-apply_analysis_fn_to_all}{}}}
\subsection{Method \code{apply_analysis_fn_to_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$apply_analysis_fn_to_all(fn = NULL, fn_name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-len"></a>}}
\if{latex}{\out{\hypertarget{method-len}{}}}
\subsection{Method \code{len()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$len()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x_seq_along"></a>}}
\if{latex}{\out{\hypertarget{method-x_seq_along}{}}}
\subsection{Method \code{x_seq_along()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$x_seq_along()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_progress"></a>}}
\if{latex}{\out{\hypertarget{method-set_progress}{}}}
\subsection{Method \code{set_progress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$set_progress(pb)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_progressor"></a>}}
\if{latex}{\out{\hypertarget{method-set_progressor}{}}}
\subsection{Method \code{set_progressor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$set_progressor(pb)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_verbose"></a>}}
\if{latex}{\out{\hypertarget{method-set_verbose}{}}}
\subsection{Method \code{set_verbose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$set_verbose(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Extracts the data provided via 'add_data' and returns it as a named list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named list, where most elements have been added via 'add_data'.

One extra named element is called 'hash'. 'hash' contains the data hashes of particular datasets/variables, as calculated using the 'spookyhash' algorithm via digest::digest.
'hash' contains two named elements:
\itemize{
\item current (the hash of the entire named list)
\item current_elements (the hash of the named elements within the named list)
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-get_analysis}{}}}
\subsection{Method \code{get_analysis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$get_analysis(index_analysis)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_argset"></a>}}
\if{latex}{\out{\hypertarget{method-get_argset}{}}}
\subsection{Method \code{get_argset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$get_argset(index_analysis)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_one_with_data"></a>}}
\if{latex}{\out{\hypertarget{method-run_one_with_data}{}}}
\subsection{Method \code{run_one_with_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$run_one_with_data(index_analysis, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_one"></a>}}
\if{latex}{\out{\hypertarget{method-run_one}{}}}
\subsection{Method \code{run_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$run_one(index_analysis, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-use_foreach_decision"></a>}}
\if{latex}{\out{\hypertarget{method-use_foreach_decision}{}}}
\subsection{Method \code{use_foreach_decision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$use_foreach_decision()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all_with_data"></a>}}
\if{latex}{\out{\hypertarget{method-run_all_with_data}{}}}
\subsection{Method \code{run_all_with_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$run_all_with_data(data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\if{latex}{\out{\hypertarget{method-run_all}{}}}
\subsection{Method \code{run_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$run_all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all_progress"></a>}}
\if{latex}{\out{\hypertarget{method-run_all_progress}{}}}
\subsection{Method \code{run_all_progress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$run_all_progress(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
