% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_opts.R
\name{set_opts}
\alias{set_opts}
\title{Set package configuration options}
\usage{
set_opts(force_verbose = FALSE)
}
\arguments{
\item{force_verbose}{Logical. Whether to force verbose output messages,
whatever the interactive state is. The default is \code{FALSE}.}
}
\value{
NULL. \code{set_opts()} changes the internal configuration of the package.
}
\description{
\code{set_opts()} sets package-wide options, such as the verbosity of output
messages. It changes the internal configuration state of the package.
}
\examples{
# Enable verbose output
set_opts(force_verbose = TRUE)

# Disable verbose output
set_opts(force_verbose = FALSE)
}
\seealso{
\link{Plan}. Its \code{verbose} argument is on by default when the session is
interactive, or when \code{force_verbose} is \code{TRUE}. See \code{vignette("plnr")} for an
introduction to the framework.
}
