% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{test_action_fn}
\alias{test_action_fn}
\title{A test action_fn for an analysis that returns the value 1}
\usage{
test_action_fn(data, argset)
}
\arguments{
\item{data}{Data}

\item{argset}{argset}
}
\description{
A test action_fn for an analysis that returns the value 1
}
