% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_again.R
\name{try_again}
\alias{try_again}
\title{try again}
\usage{
try_again(
  x,
  times = 1,
  delay_seconds_min = 1,
  delay_seconds_max = delay_seconds_min,
  verbose = FALSE
)
}
\arguments{
\item{x}{code}

\item{times}{Number of times to try}

\item{delay_seconds_min}{Number of seconds to delay on failure}

\item{delay_seconds_max}{Number of seconds to delay on failure}

\item{verbose}{Boolean. Do you want information?}
}
\description{
try again
}
