% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_again.R
\name{try_again}
\alias{try_again}
\title{Try a code snipped multiple times.}
\usage{
try_again(
  x,
  times = 2,
  delay_seconds_min = 5,
  delay_seconds_max = 10,
  verbose = FALSE
)
}
\arguments{
\item{x}{code}

\item{times}{Number of times to try}

\item{delay_seconds_min}{Number of seconds to delay on failure}

\item{delay_seconds_max}{Number of seconds to delay on failure}

\item{verbose}{Boolean. Do you want information?}
}
\description{
Adapted from function try_again from package testthat.
}
