% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{uuid_generator}
\alias{uuid_generator}
\title{Generate a UUID}
\usage{
uuid_generator()
}
\value{
A character string containing a UUID
}
\description{
Internal function to generate a unique identifier using the uuid package.
}
\keyword{internal}
