% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{uuid_generator}
\alias{uuid_generator}
\title{Generate a UUID}
\usage{
uuid_generator()
}
\value{
A character string that holds a UUID.
}
\description{
\code{uuid_generator()} is an internal function. It generates a unique identifier
with the uuid package.
}
\keyword{internal}
